grpc-gateway: switch to external annotations
Cobalt uses grpc-gateway in the servers to provide a http gateway to
grpc services. Previously, the mapping of services to http paths was
added as annotations in the proto files. However, this imposes a
dependency on "google/api/annotations.proto". Client implementations
don't need the gateway code and the dependency on annotations.proto is
thus unnecessary.

This commit switches the api mapping from in-proto annotations to an
external config file that grpc-gateway uses for code generation.